Stuck bug officially fixed in ch1
returning the colliders to 2d squares had to adjust the float point that was interfering with platforms. Created health system/ damage system. Ui is displayed. Added tile animation and better playability for Ch 1. Evil grass is fixed can jump when colliding with it. Issues that persist are enemy interact weirdly not detectable by certain angles. need to implement when player health is 0 the respawn is activated.